Rake: Track Changes to Test Files
Ensure that when any `test/corpus/*.txt` file changes
tests are run again (also re-generates the grammar as
a side effect).

Should find a better solution so that grammar generation
and tests execution are handled separately, but I'm feeling
too fatigued to come up with a better solution -- I tried,
but lack of sleep in the last two weeks is preventing me
from focusing on the task at hand.